Crate subspace_test_runtime

Modules

api

mmr

Structs

AvgBlockspaceUsageNumBlocks

BalanceTransferCheckExtension

Disable balance transfers, if configured in the runtime.

BlockAuthoringDelay

BlockHashCount

BlockHashProvider

BlockSlot

BlockSlotCount

BlockTreePruningDepth

BlockchainHistorySize

BundleLongevity

ChannelInitReservePortion

ChannelReserveFee

ConfirmationDepthK

CreditSupply

CurrentBundleAndExecutionReceiptVersion

DepositBaseFee

DepositFactor

DomainInstantiationDeposit

DomainRegistration

DomainTxRangeAdjustmentInterval

DomainsPalletId

EraDuration

ExpectedVotesPerBlock

HoldIdentifierWrapper

InitialDomainTxRange

LiquidityInfo

MaxDomainBlockSize

Use the consensus chain’s Normal extrinsics block size limit as the domain block size limit

MaxDomainBlockWeight

Use the consensus chain’s Normal extrinsics block weight limit as the domain block weight limit

MaxDomainNameLength

MaxInitialDomainAccounts

MaxOutgoingMessages

MaxPendingStakingOperation

MaxSignatories

MaximumReceiptDrift

MinInitialDomainAccountBalance

MinNominatorStake

MinOperatorStake

MinSectorLifetime

MinimumTransfer

MmrProofVerifier

MmrRootHashCount

OnChainRewards

OnChargeTransaction

Implementation of [pallet_transaction_payment::OnChargeTransaction] that charges transaction fees and distributes storage/compute fees and tip separately.

OnXDMRewards

OperatorActivationDelayInEpochs

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

PotEntropyInjectionDelay

PotEntropyInjectionInterval

PotEntropyInjectionLookbackDepth

ProposerTaxOnVotes

RecentHistoryFraction

RecentSegments

RewardPoint

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SelfChainId

SessionKeys

ShouldAdjustSolutionRange

SlotProbability

StakeEpochDuration

StakeWithdrawalLockingPeriod

StorageKeyProvider

StorageKeys

SubspaceBlockLength

We allow for 3.75 MiB for Normal extrinsic with 5 MiB maximum block length.

SubspaceBlockWeights

We allow for 2 seconds of compute with a 6 second average block time.

TotalSpacePledged

TransactionByteFee

TransactionWeightFee

TransporterEndpointId

TreasuryAccount

Version

WithdrawalLimit

Enums

AllowAuthoringBy

EnableRewardsAt

When to enable block/vote rewards

OriginCaller

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

RuntimeViewFunction

Runtime query type.

Constants

MILLISECS_PER_BLOCK

Expected block time in milliseconds.

RUNTIME_API_VERSIONS

SLOT_DURATION

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

WASM_BINARY_PATH

Type Aliases

Address

The address format for describing accounts.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Balances

BalancesConfig

Block

Block type as expected by this runtime.

Domains

DomainsConfig

Executive

Executive: handles dispatch to the various modules.

Header

Block header type as expected by this runtime.

Messenger

Mmr

Multisig

Rewards

RewardsConfig

RuntimeConfigs

RuntimeConfigsConfig

SS58Prefix

SignedExtra

The SignedExtension to the basic transaction logic.

SignedPayload

The payload being signed in transactions.

Subspace

SubspaceConfig

SubspaceMmr

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionFees

TransactionPayment

TransactionPaymentConfig

Transporter

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Utility