Crate subspace_test_runtime

Modules

api

Structs

AvgBlockspaceUsageNumBlocks

BlockAuthoringDelay

BlockHashCount

BlockHashProvider

BlockSlot

BlockSlotCount

BlockTreePruningDepth

BlockchainHistorySize

BundleLongevity

ChannelFeeModel

ChannelInitReservePortion

ChannelReserveFee

ConfirmationDepthK

CreditSupply

DomainInstantiationDeposit

DomainRegistration

DomainRuntimeUpgradeDelay

DomainTxRangeAdjustmentInterval

DomainsPalletId

EraDuration

ExpectedVotesPerBlock

InitialDomainTxRange

LiquidityInfo

MaxBundlesPerBlock

MaxDomainBlockSize

Use the consensus chain’s Normal extrinsics block size limit as the domain block size limit

MaxDomainBlockWeight

Use the consensus chain’s Normal extrinsics block weight limit as the domain block weight limit

MaxDomainNameLength

MaxInitialDomainAccounts

MaxPendingStakingOperation

MaximumReceiptDrift

MinInitialDomainAccountBalance

MinNominatorStake

MinOperatorStake

MinSectorLifetime

MinVestedTransfer

MmrProofVerifier

MmrRootHashCount

OnChainRewards

OnChargeTransaction

Implementation of [pallet_transaction_payment::OnChargeTransaction] that charges transaction fees and distributes storage/compute fees and tip separately.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

PotEntropyInjectionDelay

PotEntropyInjectionInterval

PotEntropyInjectionLookbackDepth

ProposerTaxOnVotes

RecentHistoryFraction

RecentSegments

RewardPoint

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SelfChainId

SessionKeys

ShouldAdjustSolutionRange

SlotProbability

StakeEpochDuration

StakeWithdrawalLockingPeriod

StorageKeyProvider

StorageKeys

SubspaceBlockLength

We allow for 3.75 MiB for Normal extrinsic with 5 MiB maximum block length.

SubspaceBlockWeights

We allow for 2 seconds of compute with a 6 second average block time.

TotalSpacePledged

TransactionByteFee

TransporterEndpointId

TreasuryAccount

Version

Enums

AllowAuthoringBy

EnableRewardsAt

When to enable block/vote rewards

HoldIdentifier

OriginCaller

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

Constants

DECIMAL_PLACES

Subspace Credits have 18 decimal places.

MILLISECS_PER_BLOCK

Since Subspace is probabilistic this is the average expected block time that we are targeting. Blocks will be produced at a minimum duration defined by SLOT_DURATION, but some slots will not be allocated to any farmer and hence no block will be produced. We expect to have this block time on average following the defined slot duration and the value of c configured for Subspace (where 1 - c represents the probability of a slot being empty). This value is only used indirectly to define the unit constants below that are expressed in blocks. The rest of the code should use SLOT_DURATION instead (like the Timestamp pallet for calculating the minimum period).

RUNTIME_API_VERSIONS

SHANNON

The smallest unit of the token is called Shannon.

SLOT_DURATION

SSC

One Subspace Credit.

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

Type Aliases

Address

The address format for describing accounts.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Balances

BalancesConfig

Block

Block type as expected by this runtime.

Domains

DomainsConfig

Executive

Executive: handles dispatch to the various modules.

Header

Block header type as expected by this runtime.

Messenger

Mmr

OffencesSubspace

Rewards

RewardsConfig

RuntimeConfigs

RuntimeConfigsConfig

SS58Prefix

SignedExtra

The SignedExtension to the basic transaction logic.

SignedPayload

The payload being signed in transactions.

Subspace

SubspaceConfig

SubspaceMmr

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionFees

TransactionPayment

TransactionPaymentConfig

Transporter

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Utility

Vesting

VestingConfig